Implemented Octopus Specific CI/CD
- Added support for a "permanent" pre-release of -octopus that we will use as our own internal version based off the "official" releases.

removed unused config in the main yml

re-added configure

Updated CI/CD to include build information and simplified create release config

Updated prefix to be dev so its "lower" than proper release versions (#18)

NuGet feeds order alphabetically descending after the semver, this means versions with "pr-" are considered "newer" than our  "octopus-release" branch

Update nuget package versioning to use "octopus" and "branch name" (#17)

We want to include "octopus" in the version name to make it clear that the version is a "octopus" version, and use the branch name in the version release label for PR's too.

Refactor NuGet package list in workflow file (#14)

Reverted the packages version specification as the wildcard is not supported, they were previously just defaulting to the 'latest' created package.

Fix build conflict with official package (#12)

When we 'create a release' it is specifying the package version against the "PackageId" However this causes a conflict on the package id `Nuget.CommandLine` becuse we push a new `Nuget.CommandLine` and we also consume a `Nuget.CommandLine` to do the release.

The new config will specify the new packageId and packageName only for the specific step that we have the new one, so every other instance will use the official version.