Fix build conflict with official package (#12)
When we 'create a release' it is specifying the package version against the "PackageId" However this causes a conflict on the package id `Nuget.CommandLine` becuse we push a new `Nuget.CommandLine` and we also consume a `Nuget.CommandLine` to do the release.

The new config will specify the new packageId and packageName only for the specific step that we have the new one, so every other instance will use the official version.